pyproject.toml
[tool.poetry]
name = "thirty-module"
version = "0.1.0"
description = ""
authors = ["EvgeniyRibakov <evgenijrybakov01@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django-filter = "^25.1"
django = "^4.2.7"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
djangorestframework-simplejwt = "^5.3.1"  # Исправляем имя пакета
drf-spectacular = "^0.27.2"  # Для документации
stripe = "^10.7.0"  # Для Stripe
psycopg2-binary = "^2.9.9"  # Для PostgreSQL

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"